requirements.txt
gradio
torch
torchvision
pillow
opencv-python